Move more datagen-only code to icu_provider_source
