--- components/datetime/src/provider/neo/adapter.rs
@@ -1,65 +1,0 @@
-// This file is part of ICU4X. For terms of use, please see the file
-// called LICENSE at the top level of the ICU4X source tree
-// (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
-
-use crate::provider::calendar::*;
-use crate::provider::neo::*;
-use alloc::vec;
-use icu_calendar::types::MonthCode;
-
-impl<'a> From<&months::Symbols<'a>> for MonthNames<'a> {
-    fn from(other: &months::Symbols<'a>) -> Self {
-        match other {
-            months::Symbols::SolarTwelve(cow_list) => {
-                // Can't zero-copy convert a cow list to a VarZeroVec, so we need to allocate
-                // a new VarZeroVec. Since VarZeroVec does not implement `from_iter`, first we
-                // make a Vec of string references.
-                let vec: alloc::vec::Vec<&str> = cow_list.iter().map(|x| &**x).collect();
-                MonthNames::Linear((&vec).into())
-            }
-            months::Symbols::Other(zero_map) => {
-                // Only calendar that uses this is hebrew, we can assume it is 12-month
-                let mut vec = vec![""; 24];
-
-                for (k, v) in zero_map.iter() {
-                    let Some((number, leap)) = MonthCode(*k).parsed() else {
-                        debug_assert!(false, "Found unknown month code {k}");
-                        continue;
-                    };
-                    let offset = if leap { 12 } else { 0 };
-                    if let Some(entry) = vec.get_mut((number + offset - 1) as usize) {
-                        *entry = v;
-                    } else {
-                        debug_assert!(false, "Found out of bounds hebrew month code {k}")
-                    }
-                }
-                MonthNames::LeapLinear((&vec).into())
-            }
-        }
-    }
-}
-
-impl<'a> From<&weekdays::Symbols<'a>> for LinearNames<'a> {
-    fn from(other: &weekdays::Symbols<'a>) -> Self {
-        // Input is a cow array of length 7. Need to make it a VarZeroVec.
-        let vec: alloc::vec::Vec<&str> = other.0.iter().map(|x| &**x).collect();
-        LinearNames {
-            names: (&vec).into(),
-        }
-    }
-}
-
-impl<'a> From<&day_periods::Symbols<'a>> for LinearNames<'a> {
-    fn from(other: &day_periods::Symbols<'a>) -> Self {
-        // Input is a struct with four fields. Need to make it a VarZeroVec.
-        let vec: alloc::vec::Vec<&str> = match (other.noon.as_ref(), other.midnight.as_ref()) {
-            (Some(noon), Some(midnight)) => vec![&other.am, &other.pm, &noon, &midnight],
-            (Some(noon), None) => vec![&other.am, &other.pm, &noon],
-            (None, Some(midnight)) => vec![&other.am, &other.pm, "", &midnight],
-            (None, None) => vec![&other.am, &other.pm],
-        };
-        LinearNames {
-            names: (&vec).into(),
-        }
-    }
-}
