+++ components/datetime/src/provider/pattern/hour_cycle.rs
@@ -60,7 +60,7 @@ impl CoarseHourCycle {
     #[cfg(feature = "datagen")]
     pub fn apply_on_pattern<'data>(
         &self,
-        date_time: &provider::calendar::patterns::GenericLengthPatterns<'data>,
+        date_time: &provider::skeleton::GenericLengthPatterns<'data>,
         skeletons: &provider::skeleton::DateSkeletonPatterns<'data>,
         pattern_str: &str,
         mut pattern: reference::Pattern,
