+++ components/datetime/src/provider/skeleton/mod.rs
@@ -56,13 +56,16 @@ use crate::provider::fields::components;
 
 mod available_formats;
 mod error;
+mod generic_length_patterns;
 mod helpers;
 mod plural;
 pub mod reference;
 pub mod runtime;
 #[cfg(feature = "serde")]
 mod serde;
+
 pub use available_formats::*;
 pub use error::*;
+pub use generic_length_patterns::*;
 pub use helpers::*;
 pub use plural::*;
