+++ provider/source/src/datetime/patterns.rs
@@ -3,12 +3,12 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 use crate::cldr_serde;
-use icu::datetime::provider::calendar::*;
 use icu::datetime::provider::pattern;
 use icu::datetime::provider::pattern::CoarseHourCycle;
-use icu::datetime::provider::skeleton::DateSkeletonPatterns;
+use icu::datetime::provider::skeleton::*;
+use super::legacy::*;
 
-impl From<&cldr_serde::ca::LengthPatterns> for patterns::LengthPatterns<'_> {
+impl From<&cldr_serde::ca::LengthPatterns> for LengthPatterns<'_> {
     fn from(other: &cldr_serde::ca::LengthPatterns) -> Self {
         // TODO(#308): Support numbering system variations. We currently throw them away.
         Self {
@@ -36,7 +36,7 @@ impl From<&cldr_serde::ca::LengthPatterns> for patterns::LengthPatterns<'_> {
     }
 }
 
-impl From<&cldr_serde::ca::DateTimeFormats> for patterns::LengthPatterns<'_> {
+impl From<&cldr_serde::ca::DateTimeFormats> for LengthPatterns<'_> {
     fn from(other: &cldr_serde::ca::DateTimeFormats) -> Self {
         // TODO(#308): Support numbering system variations. We currently throw them away.
         Self {
@@ -64,7 +64,7 @@ impl From<&cldr_serde::ca::DateTimeFormats> for patterns::LengthPatterns<'_> {
     }
 }
 
-impl From<&cldr_serde::ca::DateTimeFormats> for patterns::GenericLengthPatterns<'_> {
+impl From<&cldr_serde::ca::DateTimeFormats> for GenericLengthPatterns<'_> {
     fn from(other: &cldr_serde::ca::DateTimeFormats) -> Self {
         // TODO(#308): Support numbering system variations. We currently throw them away.
         Self {
@@ -94,7 +94,7 @@ impl From<&cldr_serde::ca::DateTimeFormats> for patterns::GenericLengthPatterns<
 
 impl From<&cldr_serde::ca::Dates> for DateLengths<'_> {
     fn from(other: &cldr_serde::ca::Dates) -> Self {
-        let length_combinations_v1 = patterns::GenericLengthPatterns::from(&other.datetime_formats);
+        let length_combinations_v1 = GenericLengthPatterns::from(&other.datetime_formats);
 
         Self {
             date: (&other.date_skeletons).into(),
@@ -105,7 +105,7 @@ impl From<&cldr_serde::ca::Dates> for DateLengths<'_> {
 
 impl From<&cldr_serde::ca::Dates> for TimeLengths<'_> {
     fn from(other: &cldr_serde::ca::Dates) -> Self {
-        let length_combinations_v1 = patterns::GenericLengthPatterns::from(&other.datetime_formats);
+        let length_combinations_v1 = GenericLengthPatterns::from(&other.datetime_formats);
         let skeletons_v1 = DateSkeletonPatterns::from(&other.datetime_formats.available_formats);
 
         // Note: TimeLengths is only used for preferred_hour_cycle, we don't really use
@@ -157,7 +157,7 @@ impl From<&cldr_serde::ca::Dates> for TimeLengths<'_> {
 
         let (time_h11_h12, time_h23_h24) = {
             let time = (&other.time_skeletons).into();
-            let alt_time = patterns::LengthPatterns {
+            let alt_time = LengthPatterns {
                 full: alt_hour_cycle
                     .apply_on_pattern(
                         &length_combinations_v1,
